[CREATE] put filter menu into a separate component
